% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amganalysis-package.R
\docType{package}
\name{amganalysis-package}
\alias{amganalysis}
\alias{amganalysis-package}
\title{amganalysis: Analysis of AMG Recordings in R}
\description{
Analysis of AMG recordings.
}
\author{
\strong{Maintainer}: Jesse Sadler \email{jrsadler@icloud.com} (\href{https://orcid.org/0000-0001-6081-9681}{ORCID})

}
\keyword{internal}
