% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buffer_df}
\alias{buffer_df}
\title{Create a buffer data frame to lengthen a filtered trace}
\usage{
buffer_df(df, extent, buffer, freq)
}
\arguments{
\item{df}{A trace data frame.}

\item{extent}{A vector of samples for the extent of filtered trace to which
the buffer is added. Usually this is where stimuli occur from
\code{stimuli_samples()}.}

\item{buffer}{Buffer to add to the beginning and end of the plot in
milliseconds.}

\item{freq}{Frequency of the trace recording.}
}
\description{
Used in visualization functions.
}
\keyword{internal}
