% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_basline}
\alias{check_basline}
\title{Check baseline}
\usage{
check_basline(baseline, peaks)
}
\arguments{
\item{baseline}{Baseline amplitude to be used for calculations.}

\item{peaks}{A list of numeric vectors or a numeric vectors of samples.}
}
\description{
Check that baseline is either NULL, a numeric vector of length 1, or of
length equal to the length of peaks.
}
\keyword{internal}
