% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_lengths}
\alias{check_lengths}
\title{Check that stimuli and peaks are the same length}
\usage{
check_lengths(stimuli, peaks)
}
\arguments{
\item{stimuli}{A numeric vector of stimuli from \code{stimuli_samples()}.}

\item{peaks}{A list of numeric vectors or a numeric vectors of samples.}
}
\description{
Check that there are the same number of stimuli found as peaks.
}
\keyword{internal}
