% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_samples}
\alias{check_samples}
\title{Check samples is numeric}
\usage{
check_samples(x, peaks = TRUE)
}
\arguments{
\item{x}{Object to check.}

\item{peaks}{Whether checked argument is peaks or samples.}
}
\description{
Checks that \code{x} is either a list of numeric vectors or a numeric vector.
Provides different message whether it is used to check \code{peaks} or \code{samples}.
}
\keyword{internal}
