% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_trace}
\alias{check_trace}
\title{Check trace tibble}
\usage{
check_trace(df)
}
\arguments{
\item{df}{A trace tibble to check.}
}
\description{
Check that \code{df} is a data frame. That it has variables named \code{sample},
\code{stimulus}, and \code{response}. That these are numeric.
}
\keyword{internal}
