% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{create_report}
\alias{create_report}
\title{Create AMG report}
\usage{
create_report(df, peaks, stimulus_diff = 9000, freq = 10000, baseline = NULL)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}

\item{freq}{Frequency of recording. Used to convert milliseconds to samples.}

\item{baseline}{If \code{NULL} (the default), the baseline for calculating
area under the curve is the minimum response value for each peak.
Otherwise, baseline can be a single numeric value for a consistent
baseline, or a numeric vector of length equal to the number of peaks.}
}
\value{
Returns a report of the trace in the form of a tibble with
six columns:
\itemize{
\item \code{pot_no}: The peak number.
\item \code{peak_amp}: Maximum amplitude of the peak response.
\item \code{amp_decr_pct}: Percentage decrease of the maximum amplitude from
the first peak.
\item \code{area}: Area under the curve of the response, taking into account the
\code{baseline}.
\item \code{area_decr_pct}: Percentage decrease of the area under the curve from
the first peak.
\item \code{delay_mS}: Delay between the stimulus and the onset of the response
in milliseconds.
}
}
\description{
Create AMG report
}
\examples{
x <- find_peaks_response(ex_trace_tbl)
create_report(ex_trace_tbl, x)

# Changing the baseline will affect area under the curve calculation
create_report(ex_trace_tbl, x, baseline = 0)

}
