% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_trace_tbl}
\alias{ex_trace_tbl}
\title{Example trace tibble data}
\format{
A tibble with 192,804 rows and 4 variables.
}
\usage{
ex_trace_tbl
}
\description{
A tibble derived from \code{waveMC_ex} showing the entire trace of the recording.
}
\section{Variables}{

\itemize{
\item \code{sample}: Sample number for the trace.
\item \code{secs}: Number of seconds elapsed.
\item \code{stimulus}: Reading for the stimulus.
\item \code{response}: Reading for the response.
}
}

\keyword{datasets}
