% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_waveMC}
\alias{ex_waveMC}
\title{Example waveMC data}
\format{
A waveMC object.
}
\usage{
ex_waveMC
}
\description{
waveMC object with 5 channels, but only two with data. It has a sampling
rate of 10,000 and is 19.28 seconds long, and contains 192,804 samples.
}
\keyword{datasets}
