% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-tbl.R
\name{filter_peaks}
\alias{filter_peaks}
\title{Filter out all samples that are not a peak}
\usage{
filter_peaks(df, peaks, length_out = NULL)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{length_out}{Length of each peak. Default is \code{NULL}, which leaves the
length of each peak as is. Use this argument to normalize the lengths of
the peaks so that they are all of equal length. Use \code{peaks_length()} to
check length of peaks to find good value.}
}
\value{
A tibble with two new columns: \code{sub_sample} provides the sample
number within each peak and \code{peak_nr} identifies the peaks.
}
\description{
Filter out all samples that are not a peak for purposes of plotting. The
length of each peak can either be left as is or normalized to a single
consistent length.
}
\examples{
x <- find_peaks_response(ex_trace_tbl)

filter_peaks(ex_trace_tbl, x)

# Normalize lengths of the peaks
peaks_length(x)

filter_peaks(ex_trace_tbl, x, length_out = 250)

}
