% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-peaks.R
\name{find_peaks_manual}
\alias{find_peaks_manual}
\title{Find peaks manually}
\usage{
find_peaks_manual(start, end)
}
\arguments{
\item{start}{Numeric. Where should the peak or peaks begin.}

\item{end}{Numeric. Where should the peak or peaks end.}
}
\value{
Returns a numeric vector if \code{start} and \code{end} are length 1.
Otherwise returns a list of numeric vectors.
}
\description{
Helper function to manually define peaks. This can be useful if the other
methods for finding peaks prove problematic. use visualization functions to
find where peaks should begin and end.
}
\examples{

find_peaks_manual(1, 100)

find_peaks_manual(c(1, 101, 201), c(100, 200, 300))

}
\seealso{
See \code{find_peaks_response()} and \code{find_peaks_stimulus()} for
automatic methods for defining peaks.
}
