% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-peaks.R
\name{find_peaks_response}
\alias{find_peaks_response}
\title{Find peaks by amplitude of response}
\usage{
find_peaks_response(
  df,
  baseline = 256,
  min_amp = 1500,
  min_length = 50,
  lengthen = 100
)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{baseline}{Baseline of signal. Function will find peaks above this
value using greater than. Can be vector of length 1 or length of the
number of peaks.}

\item{min_amp}{Minimum amplitude of response to find the peaks. This is
used first to find the placement of the peaks.}

\item{min_length}{Minimum number of samples that should be above \code{min_amp}
in the peaks. This helps to find the peaks and distinguish peaks
in response to stimuli from noise.}

\item{lengthen}{Number of samples to lengthen the beginning and end of the
peak from the \code{min_amp} value to reach the \code{baseline}. If this number is
too large and the \code{baseline} is too small, the peak might start before
the stimulus. If this number is too small, it might not reach the
\code{baseline}.}
}
\value{
Returns a list of the length of the number of peaks. Each list
contains the samples of each peak. These are not the recorded values
but where the peaks occur on the trace.
}
\description{
Find peaks by looking for when response is above a set \code{min_amp} for an
extended period of time (\code{min_length}). The function then finds all samples
above the \code{baseline}.
}
\seealso{
See \code{\link[=find_peaks_stimulus]{find_peaks_stimulus()}} for another way to calculate the
beginning and ending of peaks. Use \code{\link[=peaks_delay]{peaks_delay()}} to check the output.
}
