% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-peaks.R
\name{find_peaks_stimulus}
\alias{find_peaks_stimulus}
\title{Find peaks by specified delay from stimuli}
\usage{
find_peaks_stimulus(
  df,
  delay,
  freq = 10000,
  stimulus_diff = 9000,
  baseline = NULL,
  min_length = 100
)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{delay}{Delay in milliseconds from the onset of the stimulus to begin the
peak.}

\item{freq}{Frequency of recording. Used to convert milliseconds to samples.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}

\item{baseline}{Baseline of signal. Default is \code{NULL}, which chooses the
\code{baseline} for each peak based on the value of the response at the
given \code{delay} from the stimulus. Function will find peaks above this
value using greater than. Can be vector of length 1 or length of the
number of peaks.}

\item{min_length}{Minimum length of the peak in samples. This helps to
ensure measurement of peak does not end too early. If the value is too
large, the end of the peak will be below the \code{baseline}.}
}
\value{
Returns a list of numeric vectors the length of the number of peaks.
The numeric vectors correspond to the samples that distinguish each peak.
These are not the recorded values but where the peaks occur on the trace.
Use \code{find_values()} to find the response amplitudes of the returned
samples. Make sure to check that the length of the returned list is the
same as the number of stimuli. If this is not the case, arguments will
need to be changed.
}
\description{
Find peaks by finding the stimuli with \code{find_stimuli()} and then beginning
the peak a set number of milliseconds after the onset of the stimulus. The
\code{baseline} can either be defined by the user or decided separately for each
peak based on the value of the response after the given delay. This latter
method is the default.
}
\examples{

find_peaks_stimulus(ex_trace_tbl, delay = 6)

# Provide baseline
find_peaks_stimulus(ex_trace_tbl, delay = 6, baseline = 256)

}
\seealso{
See \code{\link[=find_peaks_response]{find_peaks_response()}} for another way to calculate the
beginning and ending of peaks. Use \code{\link[=peaks_delay]{peaks_delay()}} to check the output.
To manually create peaks, see \code{\link[=find_peaks_manual]{find_peaks_manual()}}.
}
