% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-stimuli.R
\name{find_stimuli}
\alias{find_stimuli}
\title{Find when stimuli were performed}
\usage{
find_stimuli(df, stimulus_diff = 9000)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{stimulus_diff}{The stimulus is found by seeing when the stimulus
recording increases by a given amount from the previous sample. Default is
9000, which seems to work for most recordings.}
}
\value{
Returns a data frame in which each row shows the onset of the
stimulus.
}
\description{
Find when stimuli were performed by change in recorded value of stimulus
column from previous sample.
}
