% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-values.R
\name{find_values}
\alias{find_values}
\title{Find values at given samples}
\usage{
find_values(df, samples)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{samples}{Numeric vector or list of vectors of sample numbers.
Usually this will be a list of samples of peaks.}
}
\value{
Returns either a list or a vector of response values depending
on whether \code{samples} is a list or vector.
}
\description{
Find values at given samples
}
