% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ms_conversion}
\alias{ms_conversion}
\alias{ms_to_samples}
\alias{samples_to_ms}
\title{Convert between milliseconds and samples at a given frequency}
\usage{
ms_to_samples(x, freq)

samples_to_ms(x, freq)
}
\arguments{
\item{x}{Value to convert.}

\item{freq}{Frequency of trace.}
}
\description{
Convert between milliseconds and samples at a given frequency
}
\keyword{internal}
