% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-analysis.R
\name{peaks_auc}
\alias{peaks_auc}
\title{Area under the curve of peaks}
\usage{
peaks_auc(df, peaks, baseline = NULL)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{baseline}{If \code{NULL} (the default), the baseline for calculating
area under the curve is the minimum response value for each peak.
Otherwise, baseline can be a single numeric value for a consistent
baseline, or a numeric vector of length equal to the number of peaks.}
}
\value{
Return a numeric vector showing the area under the curve for the
peaks.
}
\description{
Find the area under the curve for peaks of a trace tibble. Area under the
curve is calculated using trapezoidal integration provided by
\code{pracma::trapz()}. It uses the response amplitude as the \code{y} coordinate and
sample as the \code{x} coordinate.
}
\examples{
x <- find_peaks_response(ex_trace_tbl)
peaks_auc(ex_trace_tbl, x)

# Changing the baseline will affect area under the curve calculation
peaks_auc(ex_trace_tbl, x, baseline = 0)

}
