% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-checks.R
\name{peaks_checks}
\alias{peaks_checks}
\alias{peaks_min_value}
\alias{peaks_first_value}
\alias{peaks_last_value}
\title{Checks for peak groups}
\usage{
peaks_min_value(df, peaks)

peaks_first_value(df, peaks)

peaks_last_value(df, peaks)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}
}
\value{
A vector of values
}
\description{
Checks to help explore the peaks identified by \code{find_peaks_response()},
\code{find_peaks_stimuli()}, or \code{find_peaks_manual()}.
\itemize{
\item \code{peaks_min_value()} finds the minimum value of each peak, essentially
the smallest value above the baseline.
\item \code{peaks_first_value()} finds the first value of each peak. This should
correspond with, or at least be close to, the minimum and first value(s).
\item \code{peaks_last_value()} finds the last value of each peak. This should
correspond with, or at least be close to, the minimum and first value(s).
}

The values may differ slightly because the values from the recordings are
discrete and not continuous.
}
