% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-checks.R
\name{peaks_delay}
\alias{peaks_delay}
\title{Check delay between peaks and stimuli}
\usage{
peaks_delay(df, peaks, stimulus_diff = 9000, min_delay = 0, max_delay = 200)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}

\item{min_delay}{Default 0. Minimum number of samples that the peaks should
begin after the stimuli. A warning is given if the value is less.}

\item{max_delay}{Default 200. Maximum number of samples that the peaks
should begin after the stimuli. A warning is given if the value is more.}
}
\description{
This function is meant to be used to help check for any problems in finding
the peaks via \code{find_peaks_response()} or \code{find_peaks_stimulus()}. It shows
the number of samples between each stimulus and the response peak.
}
