% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-checks.R
\name{peaks_length}
\alias{peaks_length}
\title{Check length of peaks}
\usage{
peaks_length(peaks)
}
\arguments{
\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}
}
\description{
Helper function to check the number of samples in each peak. This is useful
to help with functions that normalize the length of the peaks such as
\code{filter_peaks()}.
}
