% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_list}
\alias{split_list}
\title{Split vector of values by an index into a list}
\usage{
split_list(values, index)
}
\arguments{
\item{values}{Vector of values that will be returned.}

\item{index}{Vector of index to be used to split vector into list.}
}
\description{
Creates a list whose elements are split when the index values are
non-consecutive. Useful to turn tibble to list of numeric vectors.
}
\details{
\code{values} and \code{index} will often be the same vector but need not be,
hence the use of two arguments.
}
\keyword{internal}
