% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-stimuli.R
\name{stimuli_filter}
\alias{stimuli_filter}
\title{Filter trace tibble to find onset of stimuli}
\usage{
stimuli_filter(df, stimulus_diff)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}
}
\value{
A tibble equivalent to find_stimuli() but without \code{sample_diff}
and \code{sec_diff} columns.
}
\description{
Used in \code{find_stimuli}. Making this its own function simplifies its
internal use elsewhere.
}
\keyword{internal}
