% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-stimuli.R
\name{stimuli_samples}
\alias{stimuli_samples}
\title{Find samples of stimuli}
\usage{
stimuli_samples(df, stimulus_diff)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}
}
\value{
Vector of the samples where onset of stimuli occur.
}
\description{
Simplifies the process of find_stimuli() \%>\% dplyr::pull(sample)
}
\keyword{internal}
