% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-stimuli.R
\name{stimulus_frequency}
\alias{stimulus_frequency}
\title{Find stimuli and frequency of stimuli}
\usage{
stimulus_frequency(df, stimulus_diff = 9000, sample = TRUE)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{stimulus_diff}{The stimulus is found by seeing when the stimulus
recording increases by a given amount from the previous sample. Default is
9000, which seems to work for most recordings.}

\item{sample}{Default TRUE. Should the diff column show difference in
samples or in seconds.}
}
\value{
This extends \code{find_stimuli} to show the difference between
stimuli in terms of samples or seconds.
}
\description{
Useful to find and check frequency of the stimuli by creating a \code{diff}
column to show time difference between stimuli by either sample or in
seconds.
}
\seealso{
\code{find_stimuli()}
}
