% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_peak}
\alias{viz_peak}
\title{Visualize response peak}
\usage{
viz_peak(
  df,
  peaks,
  peak_nr,
  buffer = 10,
  stimulus_diff = 9000,
  freq = 10000,
  show_peak_area = TRUE,
  highlight_color = "dodgerblue",
  baseline = NULL,
  show_stimulus = TRUE,
  annotate = TRUE,
  min_delay = 0,
  max_delay = 50
)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{peak_nr}{Number of the peak that you want to visualize.}

\item{buffer}{Buffer to add to the beginning and end of the plot in
milliseconds.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}

\item{freq}{Frequency of recording.}

\item{show_peak_area}{Whether to highlight the area of the peak.
Default is TRUE.}

\item{highlight_color}{Color to be used to highlight response peaks.}

\item{baseline}{If \code{NULL} (the default), the baseline for calculating
area under the curve and highlighting the peak area is the minimum
response value for the peak. Otherwise, baseline can be provided.}

\item{show_stimulus}{Default \code{TRUE}. Show onset of stimuli or not.}

\item{annotate}{Whether to show annotation of the area under the curve and
the delay from the onset of the stimulus to the start of the peak. Default
is TRUE.}

\item{min_delay}{Minimum delay between onset of stimulus and start of
response peak in milliseconds. If smaller, a warning is produced.}

\item{max_delay}{Maximum delay between onset of stimulus and start of
response peak in milliseconds. If larger, a warning is produced.}
}
\value{
A ggplot2 object.
}
\description{
Visualization of response peak with options to show peak area,
show stimulus, and annotate.
}
\examples{

x <- find_peaks_response(ex_trace_tbl)
viz_peak(ex_trace_tbl, x, peak_nr = 5)

# Change baseline
viz_peak(ex_trace_tbl, x, peak_nr = 5, baseline = 0)

# Change color
viz_peak(ex_trace_tbl, x, peak_nr = 5, highlight_color = "orange")

# Remove annotations and stimulus
viz_peak(ex_trace_tbl, x, peak_nr = 5,
         show_stimulus = FALSE,
         annotate = TRUE)

}
