% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_peak_facet}
\alias{viz_peak_facet}
\title{Facet wrap visualization of response peaks}
\usage{
viz_peak_facet(df, peaks)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}
}
\value{
A ggplot2 object.
}
\description{
Facet wrap visualization of response peaks
}
\examples{

x <- find_peaks_response(ex_trace_tbl)
viz_peak_facet(ex_trace_tbl, x)

}
