% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_response_interactive}
\alias{viz_response_interactive}
\title{Interactive full response trace visualization}
\usage{
viz_response_interactive(
  df,
  peaks = NULL,
  buffer = 100,
  stimulus_diff = 9000,
  freq = 10000,
  filter_response = TRUE,
  show_stimulus = TRUE,
  shade_peaks = TRUE
)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur. Default is \code{NULL}, which makes \code{shade_peaks = FALSE}.}

\item{buffer}{Buffer to add to the beginning and end of the plot in
milliseconds if \code{filter_response = TRUE}.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}

\item{freq}{Frequency of recording.}

\item{filter_response}{Whether to only visualize the response from the
first to the last stimulus (\code{TRUE}) or to visualize the whole recording
(\code{FALSE}). Used to exclude extraneous recording. Default is \code{TRUE}.}

\item{show_stimulus}{Default \code{TRUE}. Show onset of stimuli or not.}

\item{shade_peaks}{Whether to shade/highlight the area of the peaks.
Default is TRUE.}
}
\description{
Interactive visualization of a trace. The trace can be filtered from the
first stimulus to the last stimulus to better visualize peaks. Interactive
plot enables the user to zoom in on peaks. There are also options to show
onset of stimuli and shade the area of the peaks.
}
\examples{
viz_response_interactive(ex_trace_tbl)

# Shade peaks
x <- find_peaks_response(ex_trace_tbl)
viz_response_interactive(ex_trace_tbl, x)

}
\seealso{
\code{viz_response()} and \code{viz_highlight_peaks()} for ggplot2 plots.
}
