% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_stimulus}
\alias{viz_stimulus}
\title{Visualize full stimulus of trace}
\usage{
viz_stimulus(
  df,
  buffer = 100,
  stimulus_diff = 9000,
  freq = 10000,
  title = NULL,
  filter_stimulus = TRUE
)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{buffer}{Buffer to add to the beginning and end of the plot in
milliseconds if \code{filter_response = TRUE}.}

\item{stimulus_diff}{The stimulus is found by seeing when the amplitude of
the stimulus recording increases by a given amount from the previous
sample. Default is 9000, which seems to work for most recordings.}

\item{freq}{Frequency of recording.}

\item{title}{Title for the plot. A character vector of length 1.}

\item{filter_stimulus}{Whether to only visualize the stimulus from the
first to the last stimulus (\code{TRUE}) or to visualize the whole recording
(\code{FALSE}). Used to exclude extraneous recording. Default is \code{TRUE}.}
}
\description{
Visualize full stimulus of a trace. The trace can be filtered from the first
stimulus to the last stimulus to better visualize peaks or not, which may be
useful to find any oddities in the data.
}
\details{
This plot will not be of use for actual analysis. Instead, it is meant to
provide a quick and easy way to see if there are any oddities in the
stimulus data.
}
\examples{

viz_stimulus(ex_trace_tbl)

}
