% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav-to-tbl.R
\name{waveMC_to_tbl}
\alias{waveMC_to_tbl}
\title{Cast waveMC to tibble}
\usage{
waveMC_to_tbl(wav, freq)
}
\arguments{
\item{wav}{waveMC object with left and right channels.}

\item{freq}{Frequency of recording.}
}
\value{
Returns a tibble with four columns:
\itemize{
\item \code{sample}: integer representing each sample from the data.
\item \code{secs}: time value of each sample in seconds.
\item \code{stimulus}: integer reading of the stimulus recording.
\item \code{response}: integer reading of the response recording.
}

All other functions in \code{amganalysis} are built around a tibble with this
structure.
}
\description{
Cast a waveMC object created by the \code{tuneR::readWave()} to a tibble.
}
\details{
This function is based on a waveMC object with left and right
channels. The channel with the highest recorded value is made into the
stimulus variable.
}
\examples{
\dontrun{
x <- tuneR::readWave("path/to/wav/file")
waveMC_to_tbl(x, freq = 10000)
}

}
