% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-tbl.R
\name{filter_full_response}
\alias{filter_full_response}
\title{Filter data frame by response}
\usage{
filter_full_response(df, peaks, buffer = 100, freq = 10000)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{buffer}{Default is 100. Milliseconds to add before first stimulus
and after the last stimulus.}

\item{freq}{Frequency of recording. Used to convert milliseconds to samples.}
}
\description{
Filter out anything before the beginning of the first peak and after
the end of the last peak.
}
\details{
Used to plot full response.
}
\seealso{
\code{filter_full_stimuli()}
}
