% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-tbl.R
\name{filter_full_stimuli}
\alias{filter_full_stimuli}
\title{Filter data frame by stiumli}
\usage{
filter_full_stimuli(df, buffer = 100, stimulus_diff = 9000, freq = 10000)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{buffer}{Default is 100. Milliseconds to add before first stimulus
and after the last stimulus.}

\item{stimulus_diff}{The stimulus is found by seeing when the stimulus
recording increases by a given amount from the previous sample. Default is
9000, which seems to work for most recordings.}

\item{freq}{Frequency of recording. Used to convert milliseconds to samples.}
}
\description{
Filter out anything before the first stimulus and after the last stimulus.
}
\details{
Used to plot full response.
}
\seealso{
\code{filter_full_response()}
}
