% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-tbl.R
\name{filter_peaks}
\alias{filter_peaks}
\title{Filter out all samples that are not a peak}
\usage{
filter_peaks(df, peaks, length_out = 200)
}
\arguments{
\item{df}{Data frame or tibble with sample, stimulus, and response columns.}

\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{length_out}{Default is 200. The length of each peak in samples.}
}
\description{
Use to make plot with just peaks
}
