% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-checks.R
\name{peaks_length}
\alias{peaks_length}
\title{Check length of peaks}
\usage{
peaks_length(peaks)
}
\arguments{
\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}
}
\description{
Check the number of samples in each peak.
}
