% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-analysis.R
\name{peaks_length}
\alias{peaks_length}
\title{Find length of peaks}
\usage{
peaks_length(peaks, samples = TRUE, freq = NULL)
}
\arguments{
\item{peaks}{List of numeric vectors with the samples for where the peaks
occur.}

\item{samples}{Whether to return the number of samples in the peaks, the
default, or the length in milliseconds if set to \code{FALSE}.}

\item{Frequency}{of recording. Used to convert milliseconds to samples.
Default is \code{NULL} and only used if \code{samples = FALSE}.}
}
\value{
A numeric vector showing the number of samples in each peak if
\code{samples = TRUE}, or a vector showing the length of each peak in
milliseconds if \code{samples = FALSE}.
}
\description{
Find the length of peaks in either samples or milliseconds. This is useful
to help with functions that normalize the length of the peaks such as
\code{filter_peaks()}. It can also be used as part of an analysis of the peaks.
}
\examples{
x <- find_peaks_response(ex_trace_tbl)

# Find length of peaks in samples
peaks_length(x)

# Find length of peaks in milliseconds
peaks_length(x, samples = FALSE, freq = 10000)

}
